Refactoring, cleaning up
